+++ src/tap/control/setpoint/commands/unjam_integral_command.cpp
@@ -35,13 +35,6 @@ UnjamIntegralCommand::UnjamIntegralCommand(
 {
     assert(config.targetUnjamIntegralChange > 0);
     assert(config.targetCycleCount > 0);
-    assert(config.maxWaitTime > 0);
-
-    // max wait time must be > min time it will take to reach the unjam displacement given the unjam
-    // velocity
-    assert(
-        1000.0f * (this->config.targetUnjamIntegralChange / this->config.unjamSetpoint) <
-        this->config.maxWaitTime);
 
     addSubsystemRequirement(&integrableSetpointSubsystem);
 
@@ -52,8 +45,6 @@ bool UnjamIntegralCommand::isReady() { return integrableSetpointSubsystem.isOnli
 
 void UnjamIntegralCommand::initialize()
 {
-    unjamRotateTimeout.restart(config.maxWaitTime);
-
     positionBeforeUnjam = integrableSetpointSubsystem.getCurrentValueIntegral();
 
     forwardsCleared = false;
@@ -70,7 +61,7 @@ void UnjamIntegralCommand::execute()
     switch (currUnjamState)
     {
         case UNJAM_BACKWARD:
-            if (curPosition <= positionBeforeUnjam - config.targetUnjamIntegralChange)
+            if (curPosition <= integrableSetpointSubsystem.getDesiredIntegralSetpoint())
             {
                 backwardsCleared = true;
                 beginUnjamForwards();
@@ -81,7 +72,7 @@ void UnjamIntegralCommand::execute()
             }
             break;
         case UNJAM_FORWARD:
-            if (curPosition >= positionBeforeUnjam)
+            if (curPosition >= integrableSetpointSubsystem.getDesiredIntegralSetpoint())
             {
                 forwardsCleared = true;
                 beginUnjamBackwards();
@@ -119,15 +110,18 @@ bool UnjamIntegralCommand::isFinished() const
 
 void UnjamIntegralCommand::beginUnjamForwards()
 {
-    unjamRotateTimeout.restart(config.maxWaitTime);
     integrableSetpointSubsystem.setSetpoint(config.unjamSetpoint);
+    integrableSetpointSubsystem.setDesiredIntegralSetpoint(positionBeforeUnjam);
+    unjamRotateTimeout.restart(getUnjamRotateTime());
     currUnjamState = UNJAM_FORWARD;
 }
 
 void UnjamIntegralCommand::beginUnjamBackwards()
 {
-    unjamRotateTimeout.restart(config.maxWaitTime);
     integrableSetpointSubsystem.setSetpoint(-config.unjamSetpoint);
+    integrableSetpointSubsystem.setDesiredIntegralSetpoint(
+        positionBeforeUnjam - config.targetUnjamIntegralChange);
+    unjamRotateTimeout.restart(getUnjamRotateTime());
     currUnjamState = UNJAM_BACKWARD;
     backwardsCount += 1;
 }